@@ -1,48 +1,71 @@
-import 'package:flutter/foundation.dart';
-import 'package:intl/intl.dart';
-
 class Message {
-  final String message;
+  final String? message;
   final DateTime timeOfMessage;
-  final String senderID;
-  final String receiverID;
-  final String chatID;
-  final String messageType;
-  final String messageID;
+  final String? senderID;
+  final List<String>? receiverIDs;
+  final String? chatID;
+  final String? messageType;
+  final String? messageID;
+  final bool? isGroup;
 
   Message({
-    @required this.message,
-    @required this.timeOfMessage,
-    @required this.senderID,
-    @required this.chatID,
-    this.receiverID,
-    @required this.messageType,
-    @required this.messageID,
+    required this.message,
+    required this.timeOfMessage,
+    required this.senderID,
+    required this.chatID,
+    this.receiverIDs,
+    required this.messageType,
+    required this.messageID,
+    required this.isGroup,
   });
 
-  Map<String, dynamic> toMap() {
+  Map<String, dynamic> get map {
     return {
       "message": message,
       "timeSent": timeOfMessage.toString(),
       "senderID": senderID,
-      "receiverID": receiverID,
+      "receiverIDs": receiverIDs,
       "chatID": chatID,
       "messagetype": messageType,
       "messageID": messageID,
+      "isGroup": isGroup,
     };
+  }
+
+  Message copyWith({
+    String? message,
+    DateTime? timeOfMessage,
+    String? senderID,
+    List<String>? receiverIDs,
+    String? chatID,
+    String? messageType,
+    String? messageID,
+    bool? isGroup,
+  }) {
+    return Message(
+      chatID: chatID ?? this.chatID,
+      message: message ?? this.message,
+      messageID: messageID ?? this.messageID,
+      messageType: messageType ?? this.messageType,
+      senderID: senderID ?? this.senderID,
+      timeOfMessage: timeOfMessage ?? this.timeOfMessage,
+      receiverIDs: receiverIDs ?? this.receiverIDs,
+      isGroup: isGroup ?? this.isGroup,
+    );
   }
 
   Message.fromMap(Map<String, dynamic> map)
       : message = map["message"],
-        receiverID = map["receiverID"],
+        receiverIDs = List<String>.from(map["receiverIDs"]),
         senderID = map["senderID"],
         timeOfMessage = DateTime.parse(map["timeSent"] as String),
         chatID = map["chatID"],
         messageID = map["messageID"],
-        messageType = map["messagetype"];
+        messageType = map["messagetype"],
+        isGroup = map['isGroup'];
 
   @override
   String toString() {
-    return "message: $message, timeOfMessage: $timeOfMessage, senderID: $senderID, reiceverID: $receiverID , messageType: $messageType ";
+    return "message: $message, timeOfMessage: $timeOfMessage, senderID: $senderID, reiceverID: $receiverIDs , messageType: $messageType ";
   }
 }
