@@ -1,10 +1,11 @@
 import 'dart:io';
 
 import 'package:image_picker/image_picker.dart';
+import 'package:messenger/customs/error/error.dart';
 
 class MessengerImagePicker {
-  static Future<File> pickeImage() async {
-    File _imageFile;
+  static Future<File?> pickeImage() async {
+    File? _imageFile;
     try {
       var pickedFile =
           await ImagePicker().getImage(source: ImageSource.gallery);
@@ -13,6 +14,7 @@ class MessengerImagePicker {
       }
     } catch (e) {
       print(e.toString());
+      throw MessengerError(e.toString());
     }
     return _imageFile;
   }
