@@ -1,37 +1,41 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart' as firebaseAuth;
-import 'package:flutter/foundation.dart';
 import 'package:messenger/models/user.dart';
 import 'package:messenger/models/chat.dart';
 import 'package:messenger/services/online/online.dart';
 import 'package:messenger/utils/constants.dart';
 
 class FireStoreService extends Online {
-  final _cloud = FirebaseFirestore.instance;
+  final FirebaseFirestore? firebaseFirestore;
+  FireStoreService({this.firebaseFirestore})
+      : _cloud = firebaseFirestore ?? FirebaseFirestore.instance;
+  late final FirebaseFirestore _cloud;
   @override
   Future<User> saveNewUserToCloud(
-      {String userName,
-      @required String phoneNumberWithoutCC,
-      firebaseAuth.User user,
-      @required User userDataPref,
-      @required String newPhotoUrlString}) async {
+      {String? userName,
+      required String? phoneNumberWithoutCC,
+      firebaseAuth.User? user,
+      required User userDataPref,
+      required String? publicKey,
+      required String? newPhotoUrlString}) async {
     User _newUser = User(
       id: user?.uid,
       phoneNumbers: [user?.phoneNumber, phoneNumberWithoutCC],
-      photoUrl: user?.uid == userDataPref?.id && newPhotoUrlString == null
-          ? userDataPref?.photoUrl
+      photoUrl: user?.uid == userDataPref.id && newPhotoUrlString == null
+          ? userDataPref.photoUrl
           : user?.photoURL ??
               newPhotoUrlString ??
               GeneralConstants.DEFAULT_PHOTOURL,
       userName: userName,
       status: GeneralConstants.DEFAULT_STATUS,
+      publicKey: publicKey,
     );
     print(_newUser.photoUrl);
 
     await _cloud
         .collection(OnlineConstants.FIRESTORE_USER_REF)
-        .doc(_newUser?.id)
-        .set(_newUser.toMap());
+        .doc(_newUser.id!)
+        .set(_newUser.map);
     return _newUser;
   }
 
@@ -41,7 +45,7 @@ class FireStoreService extends Online {
         .collection(OnlineConstants.FIRESTORE_USER_REF)
         .doc(user.uid)
         .get();
-    return User.fromMap(_docSnapshot.data());
+    return User.fromMap(_docSnapshot.data()!);
   }
 
   @override
@@ -66,7 +70,7 @@ class FireStoreService extends Online {
         .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
         .where(
           'participantsIDs',
-          isEqualTo: query,
+          arrayContains: query,
         )
         .get();
   }
@@ -75,63 +79,136 @@ class FireStoreService extends Online {
   Future<void> createNewChat(Chat newChat) async {
     return _cloud
         .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
-        .doc(newChat.chatID)
-        .set(newChat.toMap());
+        .doc(newChat.chatID!)
+        .set(newChat.map);
   }
 
   @override
-  Stream<QuerySnapshot> getAllOnGoingchats() {
+  Future<void> deleteChat({required String id, bool isGroup = false}) async {
+    if (isGroup) {
+      return _cloud
+          .collection(OnlineConstants.FIRESTORE_ONGOING_GROUP_CHATS)
+          .doc(id)
+          .delete();
+    }
     return _cloud
         .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
-        .snapshots();
+        .doc(id)
+        .delete();
   }
 
   @override
-  Stream<QuerySnapshot> listenWhenAUserInitializesAChat(User user) {
+  Future<void> saveGroupChat(GroupChat groupChat) {
+    print("Creating...");
     return _cloud
-        .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
+        .collection(OnlineConstants.FIRESTORE_ONGOING_GROUP_CHATS)
+        .doc(groupChat.groupID)
+        .set(groupChat.map);
+  }
+
+  // @override
+  // Future<void> updateGroupChat(covariant GroupChat newGroupChat) {
+  //   return _cloud
+  //       .collection(OnlineConstants.FIRESTORE_ONGOING_GROUP_CHATS)
+  //       .doc(newGroupChat.groupID)
+  //       .set(newGroupChat.toMap());
+  // }
+
+  @override
+  Stream<QuerySnapshot> listenWhenAUserInitializesAChat(User user,
+      {bool isGroup = false}) {
+    // super.listenWhenAUserInitializesAChat(user, isGroup: isGroup);
+    return _cloud
+        .collection(isGroup
+            ? OnlineConstants.FIRESTORE_ONGOING_GROUP_CHATS
+            : OnlineConstants.FIRESTORE_ONGOING_CHATS)
         .where('participantsIDs', arrayContains: user.id)
         .snapshots();
   }
 
   @override
-  Future<bool> updateUserInCloud({User user}) async {
+  Future<bool> updateUserInCloud({User? user}) async {
     bool success = false;
     super.updateUserInCloud(user: user);
     await _cloud
         .collection(OnlineConstants.FIRESTORE_USER_REF)
-        .doc(user.id)
-        .update(user.toMap())
+        .doc(user!.id!)
+        .update(user.map)
         .then(
-      (value) async {
-        await _cloud
-            .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
-            .where('participantsIDs', arrayContains: user.id)
-            .get()
-            .then(
-          (value) async {
-            value.docs.forEach(
-              (element) async {
-                final Chat chat = Chat.froMap(element?.data());
-
-                ///checks if second user is equal to my user....incase of savedChats with self
-                final Map<String, dynamic> secondUserMap =
-                    chat.participants?.last["id"] == user.toMap()['id']
-                        ? user.toMap()
-                        : chat.participants?.last;
-                final Chat newChat = Chat(
-                    chatID: chat.chatID,
-                    participantsIDs: chat.participantsIDs,
-                    participants: [user.toMap(), secondUserMap]);
-                await element.reference.update(newChat.toMap()).then((value) {
-                  return success = true;
-                });
-              },
-            );
-          },
-        );
+      (value) {
+        _updateOnGoingChats(user).then((value) => success = true);
+        _updateOnGoingGroupChats(user).then((value) => success = true);
       },
     );
     return success;
+  }
+
+  Future<void> _updateOnGoingChats(User user) async {
+    await _cloud
+        .collection(OnlineConstants.FIRESTORE_ONGOING_CHATS)
+        .where('participantsIDs', arrayContains: user.id)
+        .get()
+        .then(
+      (value) async {
+        for (var element in value.docs) {
+          final Chat chat = Chat.froMap(element.data()!);
+          late Chat newChat;
+
+          final Map<String, dynamic>? secondUserMap = chat.participants.last!;
+
+          if (chat.participants[1]!['id'] == user.id) {
+            newChat = Chat(
+                chatID: chat.chatID,
+                participantsIDs: chat.participantsIDs,
+                participants: [chat.participants.first!, user.map]);
+          } else if (chat.participants[1]!['id'] == user.id &&
+              chat.participants[0]!['id'] == user.id) {
+            newChat = Chat(
+                chatID: chat.chatID,
+                participantsIDs: chat.participantsIDs,
+                participants: [user.map, user.map]);
+          } else {
+            newChat = Chat(
+                chatID: chat.chatID,
+                participantsIDs: chat.participantsIDs,
+                participants: [user.map, secondUserMap]);
+          }
+
+          await element.reference.update(newChat.map);
+        }
+      },
+    );
+  }
+
+  Future<void> _updateOnGoingGroupChats(User user) async {
+    await _cloud
+        .collection(OnlineConstants.FIRESTORE_ONGOING_GROUP_CHATS)
+        .where('participantsIDs', arrayContains: user.id)
+        .get()
+        .then((value) async {
+      for (var element in value.docs) {
+        final GroupChat groupChat = GroupChat.froMap(element.data()!);
+        late GroupChat newGroupChat;
+        int index = groupChat.participants
+            .indexWhere((element) => element!.containsValue(user.id));
+        int adminIndex = groupChat.groupAdmins!
+            .indexWhere((element) => element.containsValue(user.id));
+        late final User _user;
+        if (user.id == groupChat.groupCreator['id']) _user = user;
+        _user = User.fromMap(groupChat.groupCreator);
+
+        if (adminIndex != -1) {
+          groupChat.groupAdmins![adminIndex] = user.map;
+        }
+
+        groupChat.participants[index] = user.map;
+        newGroupChat = groupChat.copyWith(
+          groupCreator: _user.map,
+          participants: groupChat.participants,
+          groupAdmins: groupChat.groupAdmins,
+        );
+        await element.reference.update(newGroupChat.map);
+      }
+    });
   }
 }
