@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:messenger/customs/error/error.dart';
+import 'package:messenger/customs/custom_scroll_config.dart';
 import 'package:messenger/screens/contacts/contacts_provider.dart';
+import 'package:messenger/screens/group/group_provider.dart';
 import 'package:messenger/screens/home/home.dart';
 import 'package:messenger/services/offline/hive.db/hive_init.dart';
+import 'models/contacts_model.dart';
 import 'screens/auth/register.dart';
 import 'package:provider/provider.dart';
 import 'screens/auth/auth_provider.dart';
@@ -22,17 +26,6 @@ void main() async {
 }
 
 class MyApp extends StatelessWidget {
-  // Widget get getLaunchRoute {
-  //   if (SharedPrefs.instance.getUserData() != null) {
-  //     print("emir");
-  //     return HomeScreen();
-  //   } else {
-  //     print("Register");
-
-  //     return RegistrationScreen();
-  //   }
-  // }
-
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
@@ -40,32 +33,56 @@ class MyApp extends StatelessWidget {
         ChangeNotifierProvider<AuthProvider>(
           create: (_) => AuthProvider(),
         ),
-        ChangeNotifierProvider(create: (_) => ContactProvider()),
-        FutureProvider(
+        ChangeNotifierProvider<ContactProvider>(
+            create: (_) => ContactProvider()),
+        ChangeNotifierProvider<GroupProvider>(create: (_) => GroupProvider()),
+        FutureProvider<
+            PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts?>>(
+          initialData: PhoneContacts(
+            firstList: null,
+            lastList: null,
+          ),
           create: (_) => ContactProvider().registeredAndUnregisteredContacts(),
           catchError: (context, error) {
-            // MessengerError _messengerError = error as MessengerError;
-            // return [];
-            // ScaffoldMessenger.of(context).showSnackBar(
-            //     SnackBar(content: Text('Error fetching Contacts')));
             print(error.toString());
+            // return ;
+            throw MessengerError(error.toString());
           },
         ),
-        ChangeNotifierProvider(create: (_) => HomeProvider()),
-        ChangeNotifierProvider(create: (_) => ChatsProvider()),
-        ChangeNotifierProvider(create: (_) => ProfileProvider()),
+        ChangeNotifierProvider<HomeProvider>(create: (_) => HomeProvider()),
+        ChangeNotifierProvider<ChatsProvider>(create: (_) => ChatsProvider()),
+        ChangeNotifierProvider<ProfileProvider>(
+            create: (_) => ProfileProvider()),
+        ChangeNotifierProvider<ProfileInfoProvider>(
+            create: (_) => ProfileInfoProvider()),
       ],
       child: MaterialApp(
+        // showPerformanceOverlay: true,
         title: 'Messenger',
+        // theme: ThemeData(
+        //   scaffoldBackgroundColor: Colors.black,
+        //   brightness: Brightness.light,
+        // ),
         theme: ThemeData(
-          primarySwatch: Colors.blue,
+          brightness: Brightness.dark,
+          primarySwatch: Colors.red,
+          primaryColor: Colors.red,
         ),
-        home: SharedPrefs.instance.getUserData() != null
+        // theme: Theme.of(context),
+        home: SharedPrefs.instance.user.id != null
             ? HomeScreen()
             : RegistrationScreen(),
         builder: (context, child) {
           Utils.getBlockHeightAndWidth(context);
-          return child;
+          final MediaQueryData data = MediaQuery.of(context);
+          return ScrollConfiguration(
+            behavior: NoEffectScrollConfig(),
+            child: MediaQuery(
+              data: data.copyWith(textScaleFactor: .7),
+              child: child!,
+            ),
+          );
+          // return child!;
         },
       ),
     );
