@@ -17,18 +17,19 @@ class UserAdapter extends TypeAdapter<User> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return User(
-      id: fields[0] as String,
-      userName: fields[1] as String,
-      phoneNumbers: (fields[2] as List)?.cast<dynamic>(),
-      photoUrl: fields[3] as String,
-      status: fields[4] as String,
+      id: fields[0] as String?,
+      userName: fields[1] as String?,
+      phoneNumbers: (fields[2] as List?)?.cast<dynamic>(),
+      photoUrl: fields[3] as String?,
+      status: fields[4] as String?,
+      publicKey: fields[5] as String?,
     );
   }
 
   @override
   void write(BinaryWriter writer, User obj) {
     writer
-      ..writeByte(5)
+      ..writeByte(6)
       ..writeByte(0)
       ..write(obj.id)
       ..writeByte(1)
@@ -38,7 +39,9 @@ class UserAdapter extends TypeAdapter<User> {
       ..writeByte(3)
       ..write(obj.photoUrl)
       ..writeByte(4)
-      ..write(obj.status);
+      ..write(obj.status)
+      ..writeByte(5)
+      ..write(obj.publicKey);
   }
 
   @override
