@@ -1,6 +1,7 @@
 import 'dart:convert';
 import 'dart:ui';
 import 'package:flutter/foundation.dart';
+import 'package:messenger/app/isolate__.dart';
 import 'package:messenger/customs/error/error.dart';
 import 'package:messenger/models/contacts_model.dart';
 import 'package:messenger/models/user.dart';
@@ -13,101 +14,116 @@ import 'package:messenger/utils/constants.dart';
 import 'package:uuid/uuid.dart';
 import '../../models/chat.dart';
 
-class ContactProvider extends ChangeNotifier {
+class ContactProvider extends ChangeNotifier with ContactsIsolate_CC {
   final Online _fireStoreService = FireStoreService();
   final SharedPrefs sharedPrefs = SharedPrefs.instance;
+  bool refreshing = false;
   final _hiveHandler = HiveHandler();
-  List<List<PhoneContacts>> _listOfContact = [];
-  Future<List<List<PhoneContacts>>> registeredAndUnregisteredContacts() async {
+  PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts?>?
+      _phoneContacts;
+  Future<PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts?>>
+      registeredAndUnregisteredContacts() async {
     var _contacts = Contacts(_fireStoreService);
+    if (_phoneContacts != null) return _phoneContacts!;
     try {
-      if (!_hiveHandler.checkIfChatBoxExistAlready) {
-        await _contacts.listOfRegisteredAndUnregisteredUsers().then((value) {
-          _listOfContact = value;
-          notifyListeners();
-          _hiveHandler.saveContactsListToDB(_listOfContact);
-        });
-      } else {
-        _listOfContact = _getPhoneContactsFromHiveDB();
+      // if (_hiveHandler.checkIfChatBoxExistAlready != true) {
+      await _contacts.listOfRegisteredAndUnregisteredUsers().then((value) {
+        _phoneContacts = value;
         notifyListeners();
-      }
+        _hiveHandler.saveContactsListToDB(
+          PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts>(
+            firstList: _phoneContacts!.firstList,
+            lastList: _phoneContacts!.lastList!.map((e) => e!).toList(),
+          ),
+        );
+      });
+      // } else {
+      // _listOfContact = _getPhoneContactsFromHiveDB();
+      // notifyListeners();
+      // }
     } catch (e, s) {
       print(s.toString());
       throw MessengerError(e.toString());
     }
-    return _listOfContact;
+    return _phoneContacts!;
   }
 
   Future<void> messageUser(User myUser, User friendUser,
-      {VoidCallback navigate}) async {
+      {VoidCallback? navigate}) async {
+    print(myUser.phoneNumbers![0]);
+    print(friendUser.phoneNumbers![0]);
     Chat _chat = Chat(
       chatID: _chatID(),
       participantsIDs: [myUser.id, friendUser.id],
       participants: [
-        myUser.toMap(),
-        friendUser.toMap(),
+        myUser.map,
+        friendUser.map,
       ],
     );
     if (await _checkIfChatExistAlready(participants: _chat.participantsIDs)) {
       await _fireStoreService.createNewChat(_chat).then((value) {
         _hiveHandler.saveChatToDB(_chat).then((value) {
-          navigate();
+          navigate!();
         });
       });
     } else {
-      navigate();
+      navigate!();
     }
   }
 
-  Future<bool> _checkIfChatExistAlready({List<String> participants}) async {
-    bool exists;
-    await _fireStoreService.queryInfo(participants).then((value) {
+  Future<bool> _checkIfChatExistAlready({List<String?>? participants}) async {
+    bool exists = true;
+    await _fireStoreService.queryInfo(participants![0]).then((value) {
       bool _contains = value.docChanges.isNotEmpty;
-
-      if (_contains &&
-          value.docChanges[0].doc.data()['participantsIDs'][0] ==
-              participants[0]) {
-        exists = false;
-      } else {
-        exists = true;
-      }
+      if (_contains)
+        value.docChanges.forEach((d) {
+          if (_contains &&
+                  d.doc.data()!['participantsIDs'][0] == participants[1] ||
+              _contains &&
+                  d.doc.data()!['participantsIDs'][1] == participants[1]) {
+            exists = false;
+            print("Already Exist");
+            print(d.doc.data()!['chatID']);
+          }
+        });
     });
     return exists;
   }
 
   User getUserPref() {
     final User _user = User.fromMap(
-        json.decode(sharedPrefs.getString(OfflineConstants.MY_DATA)));
+        json.decode(sharedPrefs.getString(OfflineConstants.MY_DATA)!));
     return _user;
   }
 
-  List<List<PhoneContacts>> _getPhoneContactsFromHiveDB() {
-    List<RegisteredPhoneContacts> registered = [];
-    List<UnRegisteredPhoneContacts> unRegistered = [];
-    final _contactListFromHiveDB = _hiveHandler.getContactsListFromDB();
-    if (_contactListFromHiveDB.length > 0) {
-      print(_contactListFromHiveDB[0].length);
-      print(_contactListFromHiveDB[1].length);
-      _contactListFromHiveDB[0].forEach(
-        (element) {
-          print(element);
+  // Future<PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts>>
+  //     decodeAndCreateContactsListFromJson(
+  //         PhoneContacts<Map<String, dynamic>, Map<String, dynamic>>
+  //             list) async {
+  //   List<RegisteredPhoneContacts> registered = [];
+  //   List<UnRegisteredPhoneContacts> unRegistered = [];
+  //   return await Future<
+  //       PhoneContacts<RegisteredPhoneContacts,
+  //           UnRegisteredPhoneContacts>>.microtask(() {
+  //     for (var e in list.firstList!) {
+  //       final r = RegisteredPhoneContacts.fromMap(e);
+  //       registered.add(r);
+  //     }
+  //     for (var e in list.lastList!) {
+  //       final r = UnRegisteredPhoneContacts.fromMap(e);
+  //       unRegistered.add(r);
+  //     }
 
-          registered.add(
-            RegisteredPhoneContacts.fromMap(element),
-          );
-        },
-      );
-      _contactListFromHiveDB[1].forEach(
-        (element) {
-          unRegistered.add(
-            UnRegisteredPhoneContacts.fromMap(element),
-          );
-        },
-      );
-    } else {
-      print("List is Empty");
-    }
-    return [registered, unRegistered];
+  //     // return _listOfContact = [registered, unRegistered];
+  //     return _phoneContacts =
+  //         PhoneContacts(firstList: registered, lastList: unRegistered);
+  //   });
+  // }
+
+  Future<PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts?>?>
+      getContactsListsFromDB() async {
+    final list = _hiveHandler.getContactsListFromDB();
+    return _phoneContacts = await isolateSpawn(list);
   }
 
   String _chatID() {
@@ -115,5 +131,12 @@ class ContactProvider extends ChangeNotifier {
     return _chatID;
   }
 
-  List<List<PhoneContacts>> get listOfContact => _listOfContact;
+  dispose() {
+    disposeIsolate();
+    super.dispose();
+  }
+
+  PhoneContacts<RegisteredPhoneContacts, UnRegisteredPhoneContacts?>
+      get phoneContacts => _phoneContacts!;
+  bool get existInHive => _hiveHandler.checkIfChatBoxExistAlready;
 }
