@@ -3,7 +3,12 @@ import 'package:messenger/services/online/online.dart';
 import '../../../customs/error/error.dart';
 
 class FirebaseMAuth extends Online {
-  FirebaseAuth _auth = FirebaseAuth.instance;
+  // FirebaseAuth _auth = FirebaseAuth.instance;
+  final FirebaseAuth? firebaseAuth;
+  FirebaseMAuth({this.firebaseAuth})
+      : _auth = firebaseAuth ?? FirebaseAuth.instance;
+  late final FirebaseAuth _auth;
+
   @override
   Future<void> verifyPhoneNumber(
     phoneNumber, {
@@ -12,41 +17,52 @@ class FirebaseMAuth extends Online {
     setFirebaseUser,
     voidCallBack,
     timeOutFunction,
+    handleException,
   }) async {
-    PhoneVerificationCompleted _phoneVerificationCompleted =
-        (PhoneAuthCredential _) async {
-      try {
-        await _auth.signInWithCredential(_).then((value) {
-          fireBaseUserOnChanged().listen((user) {
-            setFirebaseUser(user);
-          });
-        }).then((value) {
-          voidCallBack();
-        });
-      } catch (e) {
-        print(e.toString());
-      }
-    };
-
-    PhoneVerificationFailed _phoneVerificationFailed =
-        (FirebaseAuthException authException) {
-      print(authException.message);
-    };
-
-    PhoneCodeSent _phoneCodeSent =
-        (String verificationId, [int forceResendingToken]) async {
-      print("::::::::::::::::" + forceResendingToken.toString());
-      print('VerifyId::::::::::::::::::: $verificationId');
-      setVerificationId(verificationId);
-    };
-
-    PhoneCodeAutoRetrievalTimeout _phoneCodeAutoRetrievalTimeout =
-        (String verificationId) {
-      print('auto Verification Timed Out');
-      timeOutFunction();
-      setPhoneAutoRetrieval(verificationId);
-    };
     try {
+      PhoneVerificationCompleted _phoneVerificationCompleted =
+          (PhoneAuthCredential _) async {
+        try {
+          await _auth.signInWithCredential(_).then((value) {
+            fireBaseUserOnChanged().listen((user) {
+              setFirebaseUser!(user!);
+            });
+          }).then((value) {
+            voidCallBack!();
+          });
+        } catch (e) {
+          print(e.toString());
+        }
+      };
+
+      PhoneVerificationFailed _phoneVerificationFailed =
+          (FirebaseAuthException authException) {
+        // print(authException.message);
+        final message = authException.code.split('-').join(' ');
+        // switch (authException.code) {
+        //   case 'invalid-phone-number':
+        //     message = 'invailid Phone Number';
+
+        //     break;
+        //   default:
+        //     message = "Something went wrong";
+        // }
+        handleException!(message);
+      };
+
+      PhoneCodeSent _phoneCodeSent =
+          (String verificationId, [int? forceResendingToken]) async {
+        print("::::::::::::::::" + forceResendingToken.toString());
+        print('VerifyId::::::::::::::::::: $verificationId');
+        setVerificationId!(verificationId, codeSent: true);
+      };
+
+      PhoneCodeAutoRetrievalTimeout _phoneCodeAutoRetrievalTimeout =
+          (String verificationId) {
+        print('auto Verification Timed Out');
+        timeOutFunction!();
+        setPhoneAutoRetrieval!(verificationId);
+      };
       await _auth.verifyPhoneNumber(
         phoneNumber: phoneNumber,
         timeout: const Duration(seconds: 20),
@@ -56,7 +72,11 @@ class FirebaseMAuth extends Online {
         codeAutoRetrievalTimeout: _phoneCodeAutoRetrievalTimeout,
       );
     } catch (e) {
-      throw MessengerError(e.toString());
+      if (e is FirebaseAuthException) {
+        throw MessengerError(e.code.split('-').join(' '));
+      } else {
+        throw MessengerError(e.toString());
+      }
     }
   }
 
@@ -66,15 +86,19 @@ class FirebaseMAuth extends Online {
   }
 
   @override
-  Future<void> verifyOTP({String verificationID, int otp}) async {
+  Future<void> verifyOTP({String? verificationID, int? otp}) async {
     try {
       final AuthCredential credential = PhoneAuthProvider.credential(
-        verificationId: verificationID,
+        verificationId: verificationID!,
         smsCode: otp.toString(),
       );
       await _auth.signInWithCredential(credential);
     } catch (e) {
-      print(e.toString());
+      if (e is FirebaseAuthException) {
+        throw MessengerError(e.code.split('-').join(' '));
+      } else {
+        throw MessengerError(e.toString());
+      }
     }
   }
 }
